split up commands + create opt in handler and request
