+++ DiscordBot/Commands/Interactive2/Graveyard/OptIn/OptInCommandRequest.cs
@@ -1,0 +1,9 @@
+using DiscordBot.Commands.Interactive2.Base.Requests;
+using DiscordBot.Common.Models.Enums;
+
+namespace DiscordBot.Commands.Interactive2.Graveyard.OptIn; 
+
+public class OptInCommandRequest : ApplicationCommandRequestBase<OptInSubCommandDefinition> {
+	public OptInCommandRequest(ApplicationCommandContext context) : base(context) { }
+	public override AuthorizationRoles MinimumAuthorizationRole => AuthorizationRoles.ClanMember;
+}
